+++ freemarker-generator-tools/src/main/java/org/apache/freemarker/generator/tools/freemarker/FreeMarkerTool.java
@@ -49,7 +49,7 @@ public class FreeMarkerTool {
      */
     public synchronized BeansWrapper getBeansWrapper() {
         if (beansWrapper == null) {
-            beansWrapper = new BeansWrapperBuilder(Configuration.VERSION_2_3_30).build();
+            beansWrapper = new BeansWrapperBuilder(Configuration.VERSION_2_3_31).build();
         }
         return beansWrapper;
     }