+++ travis.sh
@@ -19,6 +19,7 @@
 # Simulate the Travis build locally
 mvn clean install site site:stage
 cd ./freemarker-generator-cli/target/appassembler/
+sh ./bin/freemarker-generator @examples.args
 sh ./run-examples.sh
 cd ../../../freemarker-generator-maven-plugin-sample
 mvn clean package